Safari-611.1.10/LayoutTests/accessibility/mac/aria-expanded-notifications-expected.txt

Animals
Birds
This tests that when aria-expanded is toggled, the appropriate notifications are sent out.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS notifyName is "AXRowCountChanged"
PASS notifyName is "AXRowCollapsed"
PASS notifyName is "AXRowCountChanged"
PASS notifyName is "AXRowExpanded"
PASS successfullyParsed is true

TEST COMPLETE
